requirements.txt
chromadb
sentence-transformers
python-dotenv
streamlit
pandas
plotly
altair
duckdb

langgraph

langchain-google-genai
pydantic

langchain-groq
fastapi
uvicorn
python-multipart
pydantic-settings